refactor: move logic into logic.rs
